+++ Sources/TGModernNavigation/Presentation/PresentationAction.swift
@@ -5,7 +5,7 @@ import SwiftUI
 /// Modal Presentation Action
 public enum PresentationAction<R: Route>: Sendable {
     /// Present a route
-    case present(R, style: PresentationStyle, configuration: SheetConfiguration)
+    case present(R, style: PresentationStyle, configuration: SheetConfiguration, embedInNavigationStack: Bool)
     
     /// Dismiss the topmost modal
     case dismiss
@@ -17,7 +17,7 @@ public enum PresentationAction<R: Route>: Sendable {
     case dismissTo(R)
     
     /// Replace the currently presented route
-    case replace(R, style: PresentationStyle, configuration: SheetConfiguration)
+    case replace(R, style: PresentationStyle, configuration: SheetConfiguration, embedInNavigationStack: Bool)
 }
 
 extension PresentationAction: Equatable where R: Equatable {}